requirements.txt
Django==1.10
djangorestframework==3.4.7
psycopg2==2.6.2